+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0
+
+* Ability to use without a header
+* **Breaking Changes**:
+    * `AnimatedSliverExpandable` no longer performs toggling the state and instead a boolean value for `expanded` should be provided
+    * `SliverExpandableHeaderBuilder` signature changed to drop the `onToggle` callback
+
 ## 0.0.1
 
 * Initial release